% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{package}
\name{eloextend-package}
\alias{eloextend}
\alias{eloextend-package}
\title{eloextend: Extended Elo-like Ratings for Multiplayer Games}
\description{
Extends the Elo rating system to multiplayer games with 3 or more players, providing tools for rating calculation, result prediction, and visualisation and analysis of ratings histories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://john-mulvey.github.io/eloextend/}
}

}
\author{
\strong{Maintainer}: John Mulvey \email{mulveyjf@gmail.com}

}
\keyword{internal}
