% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_game_data}
\alias{validate_game_data}
\title{Validate Game Data Format}
\usage{
validate_game_data(data)
}
\arguments{
\item{data}{Data frame to validate}
}
\value{
Invisibly returns TRUE if valid, otherwise throws an error with details
}
\description{
Checks that game data is in the correct format for Elo calculations.
This function is called for its side effects (validation). It will stop
with an error if the data is invalid.
}
