% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_calculation.R
\name{get_expected_score}
\alias{get_expected_score}
\title{Calculate Expected Score for a Player}
\usage{
get_expected_score(player_rating, opponent_ratings, D = 400)
}
\arguments{
\item{player_rating}{Numeric Elo rating of the player}

\item{opponent_ratings}{Numeric vector of opponent Elo ratings}

\item{D}{Numeric scale parameter for Elo calculation. Default is 400.}
}
\value{
Numeric expected score (between 0 and 1)
}
\description{
Calculates the expected score for a player against a set of opponents
based on current Elo ratings.
}
\keyword{internal}
